Type alias for backward compatibility.
