+++ alns/alias.go
@@ -1,10 +1,16 @@
 // Package alns exports some types for backwards compatibility.
 package alns
 
-import "github.com/nextmv-io/sdk/nextroute"
+// Progressioner is an interface that can be implemented by a solver to indicate
+// that is can return the progression of the solver.
+type Progressioner interface {
+	// Progression returns the progression of the solver.
+	Progression() []ProgressionEntry
+}
 
-// ProgressionEntry in alns for backwards compatibility.
-type ProgressionEntry nextroute.ProgressionEntry
-
-// Progressioner in alns for backwards compatibility.
-type Progressioner nextroute.Progressioner
+// ProgressionEntry is a single entry in the progression of the solver.
+type ProgressionEntry struct {
+	ElapsedSeconds float64 `json:"elapsed_seconds"`
+	Value          float64 `json:"value"`
+	Iterations     int     `json:"iterations"`
+}
