+++ nextroute/solve_progressioner.go
@@ -1,15 +1,11 @@
 package nextroute
 
-// Progressioner is an interface that can be implemented by a solver to indicate
-// that is can return the progression of the solver.
-type Progressioner interface {
-	// Progression returns the progression of the solver.
-	Progression() []ProgressionEntry
-}
+import (
+	"github.com/nextmv-io/sdk/alns"
+)
 
-// ProgressionEntry is a single entry in the progression of the solver.
-type ProgressionEntry struct {
-	ElapsedSeconds float64 `json:"elapsed_seconds"`
-	Value          float64 `json:"value"`
-	Iterations     int     `json:"iterations"`
-}
+// ProgressionEntry in alns for backwards compatibility.
+type ProgressionEntry alns.ProgressionEntry
+
+// Progressioner in alns for backwards compatibility.
+type Progressioner alns.Progressioner
